refactor(UserTest) : User의 기능 일부를 Hand에 이관함에 따라 테스트 변경
